requirements.txt

asgiref==3.6.0
crispy-bootstrap4==2022.1
dj-database-url==1.2.0
Django==4.1.4
django-crispy-forms==2.0
gunicorn==20.1.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
packaging==23.0
Pillow==9.3.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-json-logger==2.0.5
python-resources==0.3
sqlparse==0.4.3


django-cors-headers==3.14.0
djangorestframework==3.14.0

python-decouple==3.8